% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Consolidate count values}
\usage{
group(
  bin,
  pattern = "*",
  fun = function(x) {
     sum(x, na.rm = T)
 },
  verbose = T
)
}
\description{
A helper function to apply a function to column names (e.g. taxa). By default, it will sum all matching column names.
}
