% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.classification.R
\name{load.classifications}
\alias{load.classifications}
\title{Load Classifications File (Plankline)}
\usage{
load.classifications(file, frame.rate = 20.5, verbose = T)
}
\description{
Load Classifications File (Plankline)
}
