% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull Classification Category}
\usage{
pull(project.dir, p = 0.5, out.dir = NULL, scratch = "/tmp/pull")
}
\arguments{
\item{project.dir}{The file path to the desired project directory}

\item{p}{The probability threshold for extracting classified images (0-1)}

\item{out.dir}{The output directory to build the file structure (default: ./tmp)}

\item{scratch}{The scratch directory to put temporary files (default: /tmp/pull)}
}
\description{
This function will extract images and generate an index file (morphocluster compatible) from a project directory.
}
\author{
Thomas Bryce Kelly
}
